Code simplification, bug fixes, and boost cases handling
1. Code simplification when computing the parabolas (curves) between an
input point and an input segment. The input segment now become the
horizontal directix used to solve the parabola equation.

2. Detection of unsolvable parabola cases. This scenario is not
completely undersood yet, but in some cases it seems that boost return
vertex that is equidistant from the input point and segment, yet the
parabola finds another point for the same x value that is also
equidistant. More investigation will happen latter on. For now, a
special exception has been created to handle that case.

3. Creation of different code units to manage rotation and distance
computation.