Implementation of bidirectional maps.
The code was using 2 maps for vertices / edges / cells. One two allow
search by object and return the index, and another one to search by
index and return an object. I implemented the use of bidirectional maps.
For 32 bits system, it enables the user to process 50 000 additional
segments and reduces the overhead for managing 2 maps.